refactor: update imports to use shared handleCommandError
